+++ src/types/emulator-options.ts
@@ -28,6 +28,7 @@ export interface EmulatorOptions {
   rom: { fileName: string; fileContent: Blob }[]
   bios: { fileName: string; fileContent: Blob }[]
   shader: { fileName: string; fileContent: Blob }[]
+  state?: Blob | undefined
 
   respondToGlobalEvents: boolean